requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
